requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==6.0.1
djangorestframework==3.16.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
uritemplate==4.2.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
